edited some formats
